feat: move expandable to shared section
