+++ src/app/Components/Artist/ArtistAbout/ArtistCareerHighlights.tsx
@@ -1,8 +1,7 @@
-import { ChevronIcon, Collapse, Flex, Text, Touchable } from "@artsy/palette-mobile"
+import { Text } from "@artsy/palette-mobile"
 import { ArtistCareerHighlights_artist$key } from "__generated__/ArtistCareerHighlights_artist.graphql"
-import { MAX_WIDTH } from "app/Components/Artist/Biography"
-import { MotiView } from "moti"
-import React, { useState } from "react"
+import { Expandable } from "app/Components/Expandable"
+import React from "react"
 import { graphql, useFragment } from "react-relay"
 
 interface ArtistCareerHighlightsProps {
@@ -51,42 +50,3 @@ const fragment = graphql`
     }
   }
 `
-
-interface ExpandableProps {
-  label?: string
-  expanded?: boolean
-  children: React.ReactNode
-}
-
-const Expandable: React.FC<ExpandableProps> = ({ children, expanded: propExpanded, label }) => {
-  const [expanded, setExpanded] = useState(propExpanded)
-
-  const handleToggle = () => {
-    setExpanded((prev) => !prev)
-  }
-
-  return (
-    <Flex borderTopWidth={1} py={1} accessibilityHint="Toggles the accordion" maxWidth={MAX_WIDTH}>
-      <Touchable
-        onPress={handleToggle}
-        accessibilityRole="togglebutton"
-        accessibilityLabel={label}
-        accessibilityState={{ expanded }}
-      >
-        <Flex flexDirection="row" alignItems="center" justifyContent="space-between">
-          <Text variant="sm">{label}</Text>
-
-          <MotiView
-            animate={{ transform: [{ rotate: !!expanded ? "-90deg" : "90deg" }] }}
-            style={{ transform: [{ rotate: !!expanded ? "-90deg" : "90deg" }] }}
-            transition={{ type: "timing" }}
-          >
-            <ChevronIcon />
-          </MotiView>
-        </Flex>
-      </Touchable>
-
-      <Collapse opened={!!expanded}>{children}</Collapse>
-    </Flex>
-  )
-}
